Try single thread with a queue
