+++ src/controller.py
@@ -12,62 +12,81 @@ import logging
 import pika
 import db as db  # type:ignore
 import settings as settings  # type:ignore
+import csv_writer as writer  # type:ignore
 
+max_threads = 1
 logging.basicConfig(format="%(levelname)s:%(asctime)s: %(message)s")
 logger = logging.getLogger(__name__)
 
-emap_db = db.starDB()
-emap_db.init_query()
-emap_db.create_connection_pool()
 
-worker_queue = queue.Queue(maxsize = 1)
+worker_queue: queue.Queue = queue.Queue(maxsize=max_threads)
 
-class waveform_message ():
+
+class waveform_message:
     def __init__(self, ch, delivery_tag, body):
         self.ch = ch
         self.delivery_tag = delivery_tag
         self.body = body
 
+
+def ack_message(ch, delivery_tag):
+    """Note that `ch` must be the same pika channel instance via which the
+    message being ACKed was retrieved (AMQP protocol constraint)."""
+    if ch.is_open:
+        ch.basic_ack(delivery_tag)
+    else:
+        logger.warning("Attempting to acknowledge a message on a closed channel.")
+
+
+def nack_message(ch, delivery_tag):
+    if ch.is_open:
+        ch.basic_nack(delivery_tag)
+    else:
+        logger.warning("Attempting to not acknowledge a message on a closed channel.")
+
+
 def waveform_callback():
-    message = worker_queue.get()
-    logger.warn(f"Got a message {message.delivery_tag}")
-    if message is not None:
-        data = json.loads(message.body)
-        location_string = data.get("mappedLocationString", "unknown")
-        observation_time = data.get("observationTime", "NaT")
-        observation_time = datetime.fromtimestamp(observation_time)
-        # I found in testing that to find the first patient I had to go back 7 months. I'm not sure this
-        # is expected, but I suppose an ICU patient could occupy a bed for a long time. Let's use
-        # 52 weeks for now.
-        start_time = observation_time - timedelta(weeks=52)
-        obs_time_str = observation_time.strftime("%Y-%m-%d:%H:%M:%S")
-        start_time_str = start_time.strftime("%Y-%m-%d:%H:%M:%S")
-        try:
-            logger.warn(f"Looking for mrn")
-            matched_mrn = emap_db.get_row(location_string, start_time_str, obs_time_str)
-        except ConnectionError:
-            cb = functools.partial(nack_message, message.ch, message.delivery_tag)
-            ch.connection.add_callback_threadsafe(cb)
-            return
+    emap_db = db.starDB()
+    emap_db.init_query()
+    emap_db.connect()
+    emap_db.create_connection_pool(5)
+    while True:
+        message = worker_queue.get()
+        if message is not None:
+            data = json.loads(message.body)
+            location_string = data.get("mappedLocationString", "unknown")
+            observation_time = data.get("observationTime", "NaT")
+            observation_time = datetime.fromtimestamp(observation_time)
+            # I found in testing that to find the first patient I had to go back 7 months. I'm not sure this
+            # is expected, but I suppose an ICU patient could occupy a bed for a long time. Let's use
+            # 52 weeks for now.
+            start_time = observation_time - timedelta(weeks=52)
+            obs_time_str = observation_time.strftime("%Y-%m-%d:%H:%M:%S")
+            start_time_str = start_time.strftime("%Y-%m-%d:%H:%M:%S")
+            try:
+                matched_mrn = emap_db.get_row(
+                    location_string, start_time_str, obs_time_str
+                )
+            except ConnectionError:
+                cb = functools.partial(nack_message, message.ch, message.delivery_tag)
+                message.ch.connection.add_callback_threadsafe(cb)
+                break
 
             if writer.write_frame(data, matched_mrn[2], matched_mrn[0]):
-                cb = functools.partial(ack_message, ch, delivery_tag)
-                ch.connection.add_callback_threadsafe(cb)
+                cb = functools.partial(ack_message, message.ch, message.delivery_tag)
+                message.ch.connection.add_callback_threadsafe(cb)
 
             worker_queue.task_done()
-    else:
-        logger.warning("empty message")
+        else:
+            logger.warning("No message in queue.")
 
 
-def on_message(ch, method_frame, _header_frame, body, args):
-    worker_queue = args
-
-    logger.warn("Got a message")
+def on_message(ch, method_frame, _header_frame, body):
     wf_message = waveform_message(ch, method_frame.delivery_tag, body)
     if not worker_queue.full():
         worker_queue.put(wf_message)
     else:
-        logger.warning("Working is queue is full.")
+        logger.warning("Working queue is full.")
 
 
 def receiver():
@@ -85,16 +104,15 @@ def receiver():
     channel.basic_qos(prefetch_count=1)
 
     threads = []
-    # I just want on thread, but in theory this should work for more 
+    # I just want on thread, but in theory this should work for more
     worker_thread = threading.Thread(target=waveform_callback)
     worker_thread.start()
     threads.append(worker_thread)
 
-    on_message_callback = functools.partial(on_message, args=(worker_queue))
     channel.basic_consume(
         queue=settings.RABBITMQ_QUEUE,
         auto_ack=False,
-        on_message_callback=on_message_callback,
+        on_message_callback=on_message,
     )
     try:
         channel.start_consuming()
